DABOM-322 feat: 미션 컴포넌트 분리
