--- apps/service/src/app/(afterLogin)/mission/log/page.tsx
@@ -1,5 +1,0 @@
-const MissionLogPage = () => {
-  return <div>mission log page</div>;
-};
-
-export default MissionLogPage;
