+++ apps/service/src/app/(afterLogin)/mission/page.tsx
@@ -1,40 +1,13 @@
 import Link from 'next/link';
 
 import { ChevronIcon } from '@icons';
-import { Button, Card, MainBox } from '@shared';
+import { Card, MainBox } from '@shared';
 
-import { MISSION_LIST_MEMBER } from 'src/data/mission';
+import MemberActionButton from 'src/components/mission/MemberActionButton';
+import OwnerActionButton from 'src/components/mission/OwnerActionButton';
+import { MISSION_LIST } from 'src/data/mission';
 
 const isOwner = true;
-
-const StatusBox = ({ children }: { children: React.ReactNode }) => (
-  <div className="text-body2-m flex h-9 w-full items-center justify-center rounded-md border border-gray-300 text-gray-700">
-    {children}
-  </div>
-);
-
-const MemberActionButton = ({ status }: { status: string }) => {
-  if (status === 'PENDING') return <StatusBox>응답 대기 중</StatusBox>;
-  return (
-    <Button size="sm" color="dark" isFullWidth>
-      보상 요청하기
-    </Button>
-  );
-};
-
-const OwnerActionButton = ({ status }: { status: string }) => {
-  if (status === 'PENDING') return <StatusBox>진행 중</StatusBox>;
-  return (
-    <div className="flex gap-2">
-      <Button size="sm" color="light" isFullWidth>
-        거절
-      </Button>
-      <Button size="sm" color="dark" isFullWidth>
-        수락
-      </Button>
-    </div>
-  );
-};
 
 const MissionPage = () => {
   return (
@@ -49,7 +22,7 @@ const MissionPage = () => {
       </p>
 
       <div className="grid grid-cols-2 gap-4 sm:grid-cols-[repeat(auto-fill,minmax(200px,1fr))]">
-        {MISSION_LIST_MEMBER.map((mission) => (
+        {MISSION_LIST.map((mission) => (
           <Card
             className="h-55"
             key={mission.id}
